Fix grenade explosion sound playing at 0, 0, 0.
Immediately deleting the grenade causes any sounds that it is emitting, including the explosion sound, to move to the center of the map, since they no longer have anywhere to emit from. Manually setting the position of the sound to the location of the grenade at the time of the explosion fixes this.